Update me crazy

+++ main.cc
@@ -28,10 +28,10 @@ int main () {
         }
         // Add a particle
         else if (choice == 2) {
-            int yIn = read("Enter the particle's x coordinate: "); //Had to swap the var name for this and the below variable
-            int xIn = read("Enter the particle's y coordinate: "); //bc the particle wouldn't print when past halfway towards the right side of the screen.
-            int dxIn = read("Enter the particle's x velocity: ");
-            int dyIn = read("Enter the particle's y velocity: ");
+            double yIn = read("Enter the particle's x coordinate: "); //Had to swap the var name for this and the below variable
+            double xIn = read("Enter the particle's y coordinate: "); //bc the particle wouldn't print when past halfway towards the right side of the screen.
+            double dxIn = read("Enter the particle's x velocity: ");
+            double dyIn = read("Enter the particle's y velocity: ");
             int lifetimeIn = read("Enter the particle's lifetime: ");
             int redIn = read("Enter the particle's red value: ");
             int greenIn = read("Enter the particle's green value: ");
